+++ Smakoowa-Api/Smakoowa-Api/Models/DatabaseModels/Statistics/RequestCount.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string ControllerName { get; set; }
         public string ActionName { get; set; }
-        public string RemainingPath { get; set; }
+        public string? RemainingPath { get; set; }
         public int Count { get; set; }
     }
 }
