+++ Smakoowa-Api/Smakoowa-Api/Services/BackgroundTaskQueue/IBackgroundTaskQueue.cs
@@ -1,0 +1,8 @@
+﻿namespace Smakoowa_Api.Services.BackgroundTaskQueue
+{
+    public interface IBackgroundTaskQueue
+    {
+        ValueTask QueueBackgroundWorkItemAsync(Func<CancellationToken, ValueTask> workItem);
+        ValueTask<Func<CancellationToken, ValueTask>> DequeueAsync(CancellationToken cancellationToken);
+    }
+}
