+++ Smakoowa-Api/Smakoowa-Api/Services/BackgroundTaskQueue/QueuedHostedService.cs
@@ -1,0 +1,46 @@
+﻿namespace Smakoowa_Api.Services.BackgroundTaskQueue
+{
+    public sealed class QueuedHostedService : BackgroundService
+    {
+        private readonly IBackgroundTaskQueue _taskQueue;
+        private readonly ILogger<QueuedHostedService> _logger;
+
+        public QueuedHostedService(IBackgroundTaskQueue taskQueue, ILogger<QueuedHostedService> logger) => (_taskQueue, _logger) = (taskQueue, logger);
+
+        protected override Task ExecuteAsync(CancellationToken stoppingToken)
+        {
+            _logger.LogInformation(
+                $"{nameof(QueuedHostedService)} is running.{Environment.NewLine}" +
+                $"{Environment.NewLine}Tap W to add a work item to the " +
+                $"background queue.{Environment.NewLine}");
+
+            return ProcessTaskQueueAsync(stoppingToken);
+        }
+
+        private async Task ProcessTaskQueueAsync(CancellationToken stoppingToken)
+        {
+            while (!stoppingToken.IsCancellationRequested)
+            {
+                try
+                {
+                    Func<CancellationToken, ValueTask>? workItem =
+                        await _taskQueue.DequeueAsync(stoppingToken);
+
+                    await workItem(stoppingToken);
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogError(ex, "Error occurred executing task work item.");
+                }
+            }
+        }
+
+        public override async Task StopAsync(CancellationToken stoppingToken)
+        {
+            _logger.LogInformation(
+                $"{nameof(QueuedHostedService)} is stopping.");
+
+            await base.StopAsync(stoppingToken);
+        }
+    }
+}
