--- Smakoowa-Api/Smakoowa-Api/Services/DefaultBackgroundTaskQueue.cs
@@ -1,171 +1,0 @@
-﻿using System.Threading.Channels;
-
-namespace Smakoowa_Api.Services
-{
-    public sealed class DefaultBackgroundTaskQueue : IBackgroundTaskQueue
-    {
-        private readonly Channel<Func<CancellationToken, ValueTask>> _queue;
-
-        public DefaultBackgroundTaskQueue(int capacity)
-        {
-            BoundedChannelOptions options = new(capacity)
-            {
-                FullMode = BoundedChannelFullMode.Wait
-            };
-            _queue = Channel.CreateBounded<Func<CancellationToken, ValueTask>>(options);
-        }
-
-        public async ValueTask QueueBackgroundWorkItemAsync(
-            Func<CancellationToken, ValueTask> workItem)
-        {
-            if (workItem is null)
-            {
-                throw new ArgumentNullException(nameof(workItem));
-            }
-
-            await _queue.Writer.WriteAsync(workItem);
-        }
-
-        public async ValueTask<Func<CancellationToken, ValueTask>> DequeueAsync(
-            CancellationToken cancellationToken)
-        {
-            Func<CancellationToken, ValueTask>? workItem =
-                await _queue.Reader.ReadAsync(cancellationToken);
-
-            return workItem;
-        }
-    }
-
-    public sealed class QueuedHostedService : BackgroundService
-    {
-        private readonly IBackgroundTaskQueue _taskQueue;
-        private readonly ILogger<QueuedHostedService> _logger;
-
-        public QueuedHostedService(
-            IBackgroundTaskQueue taskQueue,
-            ILogger<QueuedHostedService> logger) =>
-            (_taskQueue, _logger) = (taskQueue, logger);
-
-        protected override Task ExecuteAsync(CancellationToken stoppingToken)
-        {
-            _logger.LogInformation(
-                $"{nameof(QueuedHostedService)} is running.{Environment.NewLine}" +
-                $"{Environment.NewLine}Tap W to add a work item to the " +
-                $"background queue.{Environment.NewLine}");
-
-            return ProcessTaskQueueAsync(stoppingToken);
-        }
-
-        private async Task ProcessTaskQueueAsync(CancellationToken stoppingToken)
-        {
-            while (!stoppingToken.IsCancellationRequested)
-            {
-                try
-                {
-                    Func<CancellationToken, ValueTask>? workItem =
-                        await _taskQueue.DequeueAsync(stoppingToken);
-
-                    await workItem(stoppingToken);
-                }
-                catch (OperationCanceledException)
-                {
-                    // Prevent throwing if stoppingToken was signaled
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Error occurred executing task work item.");
-                }
-            }
-        }
-
-        public override async Task StopAsync(CancellationToken stoppingToken)
-        {
-            _logger.LogInformation(
-                $"{nameof(QueuedHostedService)} is stopping.");
-
-            await base.StopAsync(stoppingToken);
-        }
-    }
-
-    public sealed class MonitorLoop
-    {
-        private readonly IBackgroundTaskQueue _taskQueue;
-        private readonly ILogger<MonitorLoop> _logger;
-        private readonly CancellationToken _cancellationToken;
-
-        public MonitorLoop(
-            IBackgroundTaskQueue taskQueue,
-            ILogger<MonitorLoop> logger,
-            IHostApplicationLifetime applicationLifetime)
-        {
-            _taskQueue = taskQueue;
-            _logger = logger;
-            _cancellationToken = applicationLifetime.ApplicationStopping;
-        }
-
-        public void StartMonitorLoop()
-        {
-            _logger.LogInformation($"{nameof(MonitorAsync)} loop is starting.");
-
-            // Run a console user input loop in a background thread
-            Task.Run(async () => await MonitorAsync());
-        }
-
-        private async ValueTask MonitorAsync()
-        {
-            while (!_cancellationToken.IsCancellationRequested)
-            {
-                var keyStroke = Console.ReadKey();
-                if (keyStroke.Key == ConsoleKey.W)
-                {
-                    // Enqueue a background work item
-                    await _taskQueue.QueueBackgroundWorkItemAsync(BuildWorkItemAsync);
-                }
-            }
-        }
-
-        private async ValueTask BuildWorkItemAsync(CancellationToken token)
-        {
-            // Simulate three 5-second tasks to complete
-            // for each enqueued work item
-
-            int delayLoop = 0;
-            var guid = Guid.NewGuid();
-
-            _logger.LogInformation("Queued work item {Guid} is starting.", guid);
-
-            while (!token.IsCancellationRequested && delayLoop < 3)
-            {
-                try
-                {
-                    await Task.Delay(TimeSpan.FromSeconds(5), token);
-                }
-                catch (OperationCanceledException)
-                {
-                    // Prevent throwing if the Delay is cancelled
-                }
-
-                ++delayLoop;
-
-                _logger.LogInformation("Queued work item {Guid} is running. {DelayLoop}/3", guid, delayLoop);
-            }
-
-            string format = delayLoop switch
-            {
-                3 => "Queued Background Task {Guid} is complete.",
-                _ => "Queued Background Task {Guid} was cancelled."
-            };
-
-            _logger.LogInformation(format, guid);
-        }
-    }
-
-    public interface IBackgroundTaskQueue
-    {
-        ValueTask QueueBackgroundWorkItemAsync(
-            Func<CancellationToken, ValueTask> workItem);
-
-        ValueTask<Func<CancellationToken, ValueTask>> DequeueAsync(
-            CancellationToken cancellationToken);
-    }
-}
