--- Smakoowa-Api/Smakoowa-Api/Services/Interfaces/IBackgroundDataContext.cs
@@ -1,6 +1,0 @@
-﻿namespace Smakoowa_Api.Services.Interfaces
-{
-    public interface IBackgroundDataContext
-    {
-    }
-}
